refactor: kv store and file store of storage manager into their own module
